refactor(SE-62): wip 4/n - refactor property-related hooks
* Created a utility function to extract URL building from `usePropertySearch` to enhance modularity
* Components updated: `PropertySearchBar`
