+++ app/(home)/_components/hero-section/PropertySearchBar.tsx
@@ -15,18 +15,21 @@ interface HeroSearchBarProps {
 const PropertySearchBar: React.FC<HeroSearchBarProps> = ({ className }) => {
   const isDesktop: boolean = useMediaQuery("(min-width: 768px)");
   const { cities, isLoading, error } = usePropertyUtils();
-  const { handleSearch } = usePropertySearch();
+  const { handleRedirect } = usePropertySearch();
   const [selectedCity, setSelectedCity] = useState<string>("");
   const [budget, setBudget] = useState<string>("");
   const [date, setDate] = useState<Date | undefined>(undefined);
 
   const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    handleSearch({
-      city: selectedCity,
-      maxPrice: budget ? parseInt(budget, 10) : undefined,
-      startDate: date,
-    });
+    handleRedirect(
+      {
+        city: selectedCity,
+        maxPrice: budget ? parseInt(budget, 10) : undefined,
+        startDate: date,
+      },
+      "/properties",
+    );
   };
 
   const choices: { value: string; label: string }[] =
